refactor: move resolvers to /startup
