+++ frontend/app_flowy/lib/startup/startup.dart
@@ -2,11 +2,11 @@ import 'dart:io';
 
 import 'package:app_flowy/plugin/plugin.dart';
 import 'package:app_flowy/startup/tasks/prelude.dart';
+import 'package:app_flowy/startup/home_deps_resolver.dart';
+import 'package:app_flowy/startup/user_deps_resolver.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:app_flowy/startup/deps_resolver.dart';
-import 'package:app_flowy/user/presentation/deps_resolver.dart';
 import 'package:flowy_sdk/flowy_sdk.dart';
 
 // [[diagram: flowy startup flow]]
